pyproject.toml
[tool.poetry]
name = "grubbin"
description = "Grub's Bin-Bot"
authors = ["Franz Neulist Carroll <franzneulistcarroll@gmail.com>"]
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
cryptoxlib-aio = "^3.2.1"
asyncpg = "^0.21.0"
python-json-logger = "^0.1.11"
attrs = "^19.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.780"
black = "^19.10b0"
pytest-cov = "^2.10.0"
pre-commit = "^2.5.1"
isort = "^4.3.21"
flake8 = "^3.8.3"

[tool.poetry.scripts]
grubbin = "grubbin.console:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
